Added separate configurations for dev, prod, test. Remove typeorm.ts config file.
