+++ .test.env
@@ -1,0 +1,11 @@
+NODE_ENV=test
+APP_PORT=3000
+APP_NAME=Personal_Task_Tracker
+
+DATABASE_HOST=localhost
+DATABASE_PORT=5432
+DATABASE_USERNAME=postgres
+DATABASE_PASSWORD=password
+DATABASE_NAME=postgres
+
+JWT_SECRET_KEY=secret_for_task_tracker
